feat(#60): Correction d'une erreur dans l'ajout d'un évènemment

+++ divertech/app/src/main/java/Si3/divertech/events/EventList.java
@@ -238,7 +238,7 @@ public class EventList extends Observable {
 
         eventRef.child("title").setValue(title);
         if (pictureUrl != null) {
-            if (getEvent(eventId).getPictureUrl() != null && !getEvent(eventId).getPictureUrl().equals(pictureUrl)) {
+            if (eventId != null && getEvent(eventId).getPictureUrl() != null && !getEvent(eventId).getPictureUrl().equals(pictureUrl)) {
                 FirebaseStorage.getInstance().getReferenceFromUrl(getEvent(eventId).getPictureUrl()).delete().addOnSuccessListener(task ->
                         eventRef.child("pictureUrl").setValue(pictureUrl)
                 );
